fw-api: Update HW header files to E7 (WCSS_P3R8) for QCA6290
Update HW include header files to emulation release version E7
from E6. This specifically refers to WCSS_P3R8 folder in P4 in
FW code base. This update is just to keep the host & fw in sync.
There is almost no change to the UMAC header files that host is
concerned about in this update.

Change-Id: I2016a4d8af530971e8afce6526f960dc429ecdfa
CRs-Fixed: 2083827