UPSTREAM: sched/psi: Reduce psimon FIFO priority
PSI defaults to a FIFO-99 thread, reduce this to FIFO-1.

FIFO-99 is the very highest priority available to SCHED_FIFO and
it not a suitable default; it would indicate the psi work is the
most important work on the machine.

Since Real-Time tasks will have pre-allocated memory and locked it in
place, Real-Time tasks do not care about PSI. All it needs is to be
above OTHER.

Signed-off-by: Peter Zijlstra (Intel) <peterz@infradead.org>
Acked-by: Johannes Weiner <hannes@cmpxchg.org>
Tested-by: Suren Baghdasaryan <surenb@google.com>
Cc: Thomas Gleixner <tglx@linutronix.de>

Bug: 152499875
Test: lmkd_unit_test
Change-Id: I52964915467577bfc3543700aec9b463f6f0ffe1
(cherry picked from commit 2a220bc9f21bf95131fba36c02dd6439e6273458)
Signed-off-by: Martin Liu <liumartin@google.com>