qcacmn: Return valid channel only when pcl len > 0
In the API policy_mgr_get_alternate_channel_for_sap
the driver gets PCL for the SAP connection, but due
to various reasons such as DFS rules, SCC rules,
unsafe channels etc. the PCl can be 0.
Since there is no check of 0 PCL channels in
the same API, the driver returns the 0th element
of the PCL channel array which is case of pcl len
as 0 can be garbage, which can lead to assert if
not valid.

Fix is to validate the PCl len first and then only
return the channel to the caller API

Change-Id: Ibe4c85bb5c80026b518bf0dfeb0ce7e961286889
CRs-Fixed: 2610983