qcacmn: Update disconnect rssi on every disconnect rssi event
Currently there is a check to validate if the station is in
associated state or not in the hdd lost link cp stats info cb,
There is a race condition where if the disconnection happens
from ap in that case the callback gets invoked before the
connection state in the station context gets updated and thus
the rssi on disconnect is not updated.
To address above issue remove the check to validate the station
state and update the rssi on disconnect irrespective of the
state of the connection.

Change-Id: I9cc6a08cdf0a1081adfde7a9423a665bc977cffc
CRs-Fixed: 2685175
Moto-CR: (CR)
Reviewed-on: https://gerrit.mot.com/1554389
SLTApproved: Slta Waiver
SME-Granted: SME Approvals Granted
Tested-by: Jira Key
Reviewed-by: Ming Xiang <xiangmin@motorola.com>
Reviewed-by: Prashanth Chakravarthy <a19110@motorola.com>
Submit-Approved: Jira Key